Makefile fix
Changed target  name `gen-json` to `gen-jsonld`. Rationale:  `gen-jsonld` was referred to but not defined. `gen-json` was defined but not referred to. I assumed they might be one and the same. Fortunately for me after making this change I was able to fix an error I had (regarding `gen-jsonld` not being defined) and I was able to finally generate a linkml template in the `linkml-model-template` repository.